base: use int for ssize_t on 32-bit Windows
Use int for Perfetto's Windows ssize_t definition. This preserves
the width and value range on all Windows architectures and makes the
definition compatible with common embedders.

Electron has carried this downstream fix for Win32 builds since 2022.

+++ include/perfetto/ext/base/sys_types.h
@@ -44,7 +44,7 @@ using pid_t = int;
 #if defined(_WIN64)
 using ssize_t = int64_t;
 #else
-using ssize_t = long;
+using ssize_t = int;
 #endif  // _WIN64
 
 #endif  // __MINGW32__
